rename: RightSide 컴포넌트 디렉토리 재배치
#36
