+++ src/components/rightSide/style.tsx
